LayoutTests/platform/gtk-wk2/imported/w3c/web-platform-tests/css/css-cascade/import-conditions-expected.txt


PASS supports(display:block) is a valid import condition
PASS supports((display:flex)) is a valid import condition
PASS supports((display:block) and (display:flex)) is a valid import condition
PASS supports((display:block) and (foo:bar)) is not a valid import condition
PASS supports((display:block) or (display:flex)) is a valid import condition
PASS supports((display:block) or (foo:bar)) is a valid import condition
PASS supports(not (display: flex)) is not a valid import condition
PASS supports(display: block !important) is a valid import condition
PASS supports(foo:bar) is not a valid import condition
PASS supports(supports(display:block)) is not a valid import condition
PASS supports(()) is not a valid import condition
PASS supports() is not a valid import condition
PASS supports(display:block) (width >= 0px) is a valid import condition
PASS (width >= 0px) supports(foo:bar) is not a valid import condition
PASS (width >= 0px) supports(display:block) is not a valid import condition
PASS supports(selector(a)) is a valid import condition
PASS supports(selector(p a)) is a valid import condition
PASS supports(selector(p > a)) is a valid import condition
PASS supports(selector(p + a)) is a valid import condition
PASS supports(font-tech(color-COLRv1)) is a valid import condition
PASS supports(font-tech(invalid)) is not a valid import condition
PASS supports(font-format(opentype)) is a valid import condition
PASS supports(font-format(woff)) is a valid import condition
PASS supports(font-format(invalid)) is not a valid import condition
PASS layer(A.B) supports(font-format(opentype)) is a valid import condition
PASS layer supports(selector(a)) is a valid import condition
